started product work in order settings

+++ database/migrations/2024_05_23_093006_drop_and_rename_column_from_inventories_table.php
@@ -1,8 +1,9 @@
 <?php
 
-use Illuminate\Database\Migrations\Migration;
-use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
 
 return new class extends Migration
 {
@@ -12,7 +13,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('inventories', function (Blueprint $table) {
-            $table->dropColumn('stockin_date');
+            if (Schema::hasColumn('inventories', 'stockin_date')) {
+                $table->dropColumn('stockin_date');
+            }
             $table->dropColumn('purchase_order_no');
             $table->dropColumn('purchase_order_date');
             $table->dropColumn('purchase_order_amount');
@@ -24,7 +27,7 @@ return new class extends Migration
             $table->dropColumn('narration');
             $table->renameColumn('quantity', 'opening_stock');
             $table->renameColumn('low_atock', 'low_stock');
-            \DB::statement("ALTER TABLE `inventories` CHANGE `inventory_type` `inventory_type` ENUM('opening', 'manual', 'automatic') NOT NULL DEFAULT 'opening';");
+            DB::statement("ALTER TABLE `inventories` CHANGE `inventory_type` `inventory_type` ENUM('opening', 'manual', 'automatic') NOT NULL DEFAULT 'opening';");
         });
     }
 
@@ -44,8 +47,8 @@ return new class extends Migration
             $table->dropColumn('file');
             $table->dropColumn('vendor_id');
             $table->dropColumn('narration');
-            $table->renameColumn('quantity', 'opening_stock');
-            $table->renameColumn('low_atock', 'low_stock');
+            $table->renameColumn('opening_stock', 'quantity');
+            $table->renameColumn('low_stock', 'low_atock');
         });
     }
 };
